Move password generation password to common config
