+++ config/rds/config.go
@@ -5,29 +5,14 @@ Copyright 2021 Upbound Inc.
 package rds
 
 import (
-	"context"
 	"fmt"
 
-	v1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
-	"github.com/crossplane/crossplane-runtime/pkg/errors"
-	"github.com/crossplane/crossplane-runtime/pkg/fieldpath"
-	"github.com/crossplane/crossplane-runtime/pkg/meta"
-	"github.com/crossplane/crossplane-runtime/pkg/password"
-	"github.com/crossplane/crossplane-runtime/pkg/reconciler/managed"
-	"github.com/crossplane/crossplane-runtime/pkg/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/crossplane/upjet/pkg/config"
 	"github.com/crossplane/upjet/pkg/types/comments"
 
 	"github.com/upbound/provider-aws/config/common"
-)
-
-const (
-	errGetPasswordSecret = "cannot get password secret"
 )
 
 // Configure adds configurations for the rds group.
@@ -135,7 +120,7 @@ func Configure(p *config.Provider) {
 			Description: desc.String(),
 		}
 		r.InitializerFns = append(r.InitializerFns,
-			PasswordGenerator(
+			common.PasswordGenerator(
 				"spec.forProvider.passwordSecretRef",
 				"spec.forProvider.autoGeneratePassword",
 			))
@@ -190,59 +175,4 @@ func Configure(p *config.Provider) {
 	p.AddResourceConfigurator("aws_db_cluster_snapshot", func(r *config.Resource) {
 		r.UseAsync = true
 	})
-}
-
-// PasswordGenerator returns an InitializerFn that will generate a password
-// for a resource if the toggle field is set to true and the secret referenced
-// by the secretRefFieldPath is not found or does not have content corresponding
-// to the password key.
-func PasswordGenerator(secretRefFieldPath, toggleFieldPath string) config.NewInitializerFn { //nolint:gocyclo
-	// NOTE(muvaf): This function is just 1 point over the cyclo limit but there
-	// is no easy way to reduce it without making it harder to read.
-	return func(client client.Client) managed.Initializer {
-		return managed.InitializerFn(func(ctx context.Context, mg resource.Managed) error {
-			paved, err := fieldpath.PaveObject(mg)
-			if err != nil {
-				return errors.Wrap(err, "cannot pave object")
-			}
-			sel := &v1.SecretKeySelector{}
-			if err := paved.GetValueInto(secretRefFieldPath, sel); err != nil {
-				return errors.Wrapf(resource.Ignore(fieldpath.IsNotFound, err), "cannot unmarshal %s into a secret key selector", secretRefFieldPath)
-			}
-			s := &corev1.Secret{}
-			if err := client.Get(ctx, types.NamespacedName{Namespace: sel.Namespace, Name: sel.Name}, s); resource.IgnoreNotFound(err) != nil {
-				return errors.Wrap(err, errGetPasswordSecret)
-			}
-			if err == nil && len(s.Data[sel.Key]) != 0 {
-				// Password is already set.
-				return nil
-			}
-			// At this point, either the secret doesn't exist, or it doesn't
-			// have the password filled.
-			if gen, err := paved.GetBool(toggleFieldPath); err != nil || !gen {
-				// If there is error, then we return that.
-				// If the toggle field is not set to true, then we return nil.
-				// Because we don't want to generate a password if the user
-				// doesn't want to.
-				return errors.Wrapf(resource.Ignore(fieldpath.IsNotFound, err), "cannot get the value of %s", toggleFieldPath)
-			}
-			pw, err := password.Generate()
-			if err != nil {
-				return errors.Wrap(err, "cannot generate password")
-			}
-			s.SetName(sel.Name)
-			s.SetNamespace(sel.Namespace)
-			if !meta.WasCreated(s) {
-				// We don't want to own the Secret if it is created by someone
-				// else, otherwise the deletion of the managed resource will
-				// delete the Secret that we didn't create in the first place.
-				meta.AddOwnerReference(s, meta.AsOwner(meta.TypedReferenceTo(mg, mg.GetObjectKind().GroupVersionKind())))
-			}
-			if s.Data == nil {
-				s.Data = make(map[string][]byte, 1)
-			}
-			s.Data[sel.Key] = []byte(pw)
-			return errors.Wrap(resource.NewAPIPatchingApplicator(client).Apply(ctx, s), "cannot apply password secret")
-		})
-	}
 }
